species specific sprites
I've set up the framework for species specific sprites.

You can find the start of this implementation in the icons/mob/uniform file. I put a "breaker" in between the plasmamen sprites and the regular human sprites, and between the vox sprites and the plasmamen sprites.

There's still a lot of sprites that need to be ported, but at this point the end is in sight when it comes to porting over the major organic species, there's no major features that need to be implemented beyond.